This ensures we are using the latest code from Log4Net.Async for these libs. there are very little changes and none breaking but it is now up-to-date and we will use the nuget package in v8 instead.
