+++ src/Umbraco.Core/Logging/LoggingEventContext.cs
@@ -3,15 +3,18 @@ using log4net.Core;
 namespace Umbraco.Core.Logging
 {
     /// <remarks>
-    /// Based on https://github.com/cjbhaines/Log4Net.Async
+    /// Borrowed from https://github.com/cjbhaines/Log4Net.Async - will reference Nuget packages directly in v8
     /// </remarks>
-    internal class LoggingEventContext
+    internal sealed class LoggingEventContext
     {
-        public LoggingEventContext(LoggingEvent loggingEvent)
+        public LoggingEventContext(LoggingEvent loggingEvent, object httpContext)
         {
             LoggingEvent = loggingEvent;
+            HttpContext = httpContext;
         }
 
         public LoggingEvent LoggingEvent { get; set; }
+
+        public object HttpContext { get; set; }
     }
 }
