+++ src/Umbraco.Tests/Logging/DebugAppender.cs
@@ -5,17 +5,16 @@ using log4net.Core;
 
 namespace Umbraco.Tests.Logging
 {
-    internal class DebugAppender : MemoryAppender
+    /// <summary>
+    /// Borrowed from https://github.com/cjbhaines/Log4Net.Async - will reference Nuget packages directly in v8
+    /// </summary>
+    public class DebugAppender : MemoryAppender
     {
         public TimeSpan AppendDelay { get; set; }
         public int LoggedEventCount { get { return m_eventsList.Count; } }
 
-        public bool Cancel { get; set; }
-
         protected override void Append(LoggingEvent loggingEvent)
         {
-            if (Cancel) return;
-
             if (AppendDelay > TimeSpan.Zero)
             {
                 Thread.Sleep(AppendDelay);
